Quality of Life changes + Blackwind map+movement
a lot went down this update. there is a temp package representing
possible ways to handle player movement and mapping. a map editor class
is in these files as well, saving maps as plaintext tileID's.

as for the QoL's, i removed the need to import a JFrame into Battle and
PauseMenu, as it was redundant and almost useless. more to come :)